refactor: instance 폴더 구조 변경
